+++ analysis/flip_measurement/run_flip.py
@@ -8,7 +8,7 @@ import argparse
 import numpy as np
 from coffea import processor
 from coffea.nanoevents import NanoAODSchema
-import topeftenv
+import topcoffea.modules.topeftenv
 
 import flip_mr_processor
